new case studies added
