--- ac1/htmlcov/d_a44f0ac069e85531_test_usr_3_course_score_py.html
@@ -1,106 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <title>Coverage for tests\test_usr_3_course_score.py: 100%</title>
-    <link rel="icon" sizes="32x32" href="favicon_32.png">
-    <link rel="stylesheet" href="style.css" type="text/css">
-    <script type="text/javascript" src="coverage_html.js" defer></script>
-</head>
-<body class="pyfile">
-<header>
-    <div class="content">
-        <h1>
-            <span class="text">Coverage for </span><b>tests\test_usr_3_course_score.py</b>:
-            <span class="pc_cov">100%</span>
-        </h1>
-        <aside id="help_panel_wrapper">
-            <input id="help_panel_state" type="checkbox">
-            <label for="help_panel_state">
-                <img id="keyboard_icon" src="keybd_closed.png" alt="Show/hide keyboard shortcuts" />
-            </label>
-            <div id="help_panel">
-                <p class="legend">Shortcuts on this page</p>
-                <div class="keyhelp">
-                    <p>
-                        <kbd>r</kbd>
-                        <kbd>m</kbd>
-                        <kbd>x</kbd>
-                        &nbsp; toggle line displays
-                    </p>
-                    <p>
-                        <kbd>j</kbd>
-                        <kbd>k</kbd>
-                        &nbsp; next/prev highlighted chunk
-                    </p>
-                    <p>
-                        <kbd>0</kbd> &nbsp; (zero) top of page
-                    </p>
-                    <p>
-                        <kbd>1</kbd> &nbsp; (one) first highlighted chunk
-                    </p>
-                    <p>
-                        <kbd>[</kbd>
-                        <kbd>]</kbd>
-                        &nbsp; prev/next file
-                    </p>
-                    <p>
-                        <kbd>u</kbd> &nbsp; up to the index
-                    </p>
-                    <p>
-                        <kbd>?</kbd> &nbsp; show/hide this help
-                    </p>
-                </div>
-            </div>
-        </aside>
-        <h2>
-            <span class="text">5 statements &nbsp;</span>
-            <button type="button" class="run button_toggle_run" value="run" data-shortcut="r" title="Toggle lines run">5<span class="text"> run</span></button>
-            <button type="button" class="mis show_mis button_toggle_mis" value="mis" data-shortcut="m" title="Toggle lines missing">0<span class="text"> missing</span></button>
-            <button type="button" class="exc show_exc button_toggle_exc" value="exc" data-shortcut="x" title="Toggle lines excluded">0<span class="text"> excluded</span></button>
-        </h2>
-        <p class="text">
-            <a id="prevFileLink" class="nav" href="d_a44f0ac069e85531_test_usr_2_premium_users_py.html">&#xab; prev</a> &nbsp; &nbsp;
-            <a id="indexLink" class="nav" href="index.html">&Hat; index</a> &nbsp; &nbsp;
-            <a id="nextFileLink" class="nav" href="d_a44f0ac069e85531_test_usr_3_fail_py.html">&#xbb; next</a>
-            &nbsp; &nbsp; &nbsp;
-            <a class="nav" href="https://coverage.readthedocs.io/en/7.3.1">coverage.py v7.3.1</a>,
-            created at 2023-09-25 00:31 -0300
-        </p>
-        <aside class="hidden">
-            <button type="button" class="button_next_chunk" data-shortcut="j"/>
-            <button type="button" class="button_prev_chunk" data-shortcut="k"/>
-            <button type="button" class="button_top_of_page" data-shortcut="0"/>
-            <button type="button" class="button_first_chunk" data-shortcut="1"/>
-            <button type="button" class="button_prev_file" data-shortcut="["/>
-            <button type="button" class="button_next_file" data-shortcut="]"/>
-            <button type="button" class="button_to_index" data-shortcut="u"/>
-            <button type="button" class="button_show_hide_help" data-shortcut="?"/>
-        </aside>
-    </div>
-</header>
-<main id="source">
-    <p class="run"><span class="n"><a id="t1" href="#t1">1</a></span><span class="t"><span class="key">from</span> <span class="nam">app</span><span class="op">.</span><span class="nam">game_plat</span> <span class="key">import</span> <span class="nam">CoursePlatform</span>&nbsp;</span><span class="r"></span></p>
-    <p class="pln"><span class="n"><a id="t2" href="#t2">2</a></span><span class="t">&nbsp;</span><span class="r"></span></p>
-    <p class="pln"><span class="n"><a id="t3" href="#t3">3</a></span><span class="t">&nbsp;</span><span class="r"></span></p>
-    <p class="run"><span class="n"><a id="t4" href="#t4">4</a></span><span class="t"><span class="key">def</span> <span class="nam">testPassScoreToCourse</span><span class="op">(</span><span class="op">)</span><span class="op">:</span>&nbsp;</span><span class="r"></span></p>
-    <p class="run"><span class="n"><a id="t5" href="#t5">5</a></span><span class="t">    <span class="nam">platform</span> <span class="op">=</span> <span class="nam">CoursePlatform</span><span class="op">(</span><span class="op">)</span>&nbsp;</span><span class="r"></span></p>
-    <p class="run"><span class="n"><a id="t6" href="#t6">6</a></span><span class="t">    <span class="nam">platform</span><span class="op">.</span><span class="nam">RegisterUser</span><span class="op">(</span><span class="op">)</span>&nbsp;</span><span class="r"></span></p>
-    <p class="run"><span class="n"><a id="t7" href="#t7">7</a></span><span class="t">    <span class="key">assert</span> <span class="nam">platform</span><span class="op">.</span><span class="nam">ScoreToCourse</span><span class="op">(</span>&nbsp;</span><span class="r"></span></p>
-    <p class="pln"><span class="n"><a id="t8" href="#t8">8</a></span><span class="t">        <span class="nam">ruler_number</span><span class="op">=</span><span class="num">12</span><span class="op">)</span><span class="op">,</span> <span class="str">"O Usu&#225;rio n&#227;o possui curso ou a quantidade \</span>&nbsp;</span><span class="r"></span></p>
-    <p class="pln"><span class="n"><a id="t9" href="#t9">9</a></span><span class="t"><span class="str">            de cursos &#233; superior/inferior que o ruler_number"</span>&nbsp;</span><span class="r"></span></p>
-</main>
-<footer>
-    <div class="content">
-        <p>
-            <a id="prevFileLink" class="nav" href="d_a44f0ac069e85531_test_usr_2_premium_users_py.html">&#xab; prev</a> &nbsp; &nbsp;
-            <a id="indexLink" class="nav" href="index.html">&Hat; index</a> &nbsp; &nbsp;
-            <a id="nextFileLink" class="nav" href="d_a44f0ac069e85531_test_usr_3_fail_py.html">&#xbb; next</a>
-            &nbsp; &nbsp; &nbsp;
-            <a class="nav" href="https://coverage.readthedocs.io/en/7.3.1">coverage.py v7.3.1</a>,
-            created at 2023-09-25 00:31 -0300
-        </p>
-    </div>
-</footer>
-</body>
-</html>
