--- ac1/htmlcov/index.html
@@ -1,172 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <title>Coverage report</title>
-    <link rel="icon" sizes="32x32" href="favicon_32.png">
-    <link rel="stylesheet" href="style.css" type="text/css">
-    <script type="text/javascript" src="coverage_html.js" defer></script>
-</head>
-<body class="indexfile">
-<header>
-    <div class="content">
-        <h1>Coverage report:
-            <span class="pc_cov">95%</span>
-        </h1>
-        <aside id="help_panel_wrapper">
-            <input id="help_panel_state" type="checkbox">
-            <label for="help_panel_state">
-                <img id="keyboard_icon" src="keybd_closed.png" alt="Show/hide keyboard shortcuts" />
-            </label>
-            <div id="help_panel">
-                <p class="legend">Shortcuts on this page</p>
-                <div class="keyhelp">
-                    <p>
-                        <kbd>n</kbd>
-                        <kbd>s</kbd>
-                        <kbd>m</kbd>
-                        <kbd>x</kbd>
-                        <kbd>c</kbd>
-                        &nbsp; change column sorting
-                    </p>
-                    <p>
-                        <kbd>[</kbd>
-                        <kbd>]</kbd>
-                        &nbsp; prev/next file
-                    </p>
-                    <p>
-                        <kbd>?</kbd> &nbsp; show/hide this help
-                    </p>
-                </div>
-            </div>
-        </aside>
-        <form id="filter_container">
-            <input id="filter" type="text" value="" placeholder="filter..." />
-        </form>
-        <p class="text">
-            <a class="nav" href="https://coverage.readthedocs.io/en/7.3.1">coverage.py v7.3.1</a>,
-            created at 2023-09-25 00:31 -0300
-        </p>
-    </div>
-</header>
-<main id="index">
-    <table class="index" data-sortable>
-        <thead>
-            <tr class="tablehead" title="Click to sort">
-                <th class="name left" aria-sort="none" data-shortcut="n">Module</th>
-                <th aria-sort="none" data-default-sort-order="descending" data-shortcut="s">statements</th>
-                <th aria-sort="none" data-default-sort-order="descending" data-shortcut="m">missing</th>
-                <th aria-sort="none" data-default-sort-order="descending" data-shortcut="x">excluded</th>
-                <th class="right" aria-sort="none" data-shortcut="c">coverage</th>
-            </tr>
-        </thead>
-        <tbody>
-            <tr class="file">
-                <td class="name left"><a href="d_5f5a17c013354698___init___py.html">app\__init__.py</a></td>
-                <td>0</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="0 0">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_5f5a17c013354698_game_plat_py.html">app\game_plat.py</a></td>
-                <td>56</td>
-                <td>4</td>
-                <td>0</td>
-                <td class="right" data-ratio="52 56">93%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531___init___py.html">tests\__init__.py</a></td>
-                <td>0</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="0 0">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_1_fail_py.html">tests\test_usr_1_fail.py</a></td>
-                <td>2</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="2 2">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_1_score_to_cripto_py.html">tests\test_usr_1_score_to_cripto.py</a></td>
-                <td>4</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="4 4">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_2_fail_py.html">tests\test_usr_2_fail.py</a></td>
-                <td>2</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="2 2">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_2_premium_users_py.html">tests\test_usr_2_premium_users.py</a></td>
-                <td>5</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="5 5">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_3_course_score_py.html">tests\test_usr_3_course_score.py</a></td>
-                <td>5</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="5 5">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_3_fail_py.html">tests\test_usr_3_fail.py</a></td>
-                <td>2</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="2 2">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_4_course_bonus_py.html">tests\test_usr_4_course_bonus.py</a></td>
-                <td>9</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="9 9">100%</td>
-            </tr>
-            <tr class="file">
-                <td class="name left"><a href="d_a44f0ac069e85531_test_usr_4_fail_py.html">tests\test_usr_4_fail.py</a></td>
-                <td>2</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="2 2">100%</td>
-            </tr>
-        </tbody>
-        <tfoot>
-            <tr class="total">
-                <td class="name left">Total</td>
-                <td>87</td>
-                <td>4</td>
-                <td>0</td>
-                <td class="right" data-ratio="83 87">95%</td>
-            </tr>
-        </tfoot>
-    </table>
-    <p id="no_rows">
-        No items found using the specified filter.
-    </p>
-</main>
-<footer>
-    <div class="content">
-        <p>
-            <a class="nav" href="https://coverage.readthedocs.io/en/7.3.1">coverage.py v7.3.1</a>,
-            created at 2023-09-25 00:31 -0300
-        </p>
-    </div>
-    <aside class="hidden">
-        <a id="prevFileLink" class="nav" href="d_a44f0ac069e85531_test_usr_4_fail_py.html"/>
-        <a id="nextFileLink" class="nav" href="d_5f5a17c013354698___init___py.html"/>
-        <button type="button" class="button_prev_file" data-shortcut="["/>
-        <button type="button" class="button_next_file" data-shortcut="]"/>
-        <button type="button" class="button_show_hide_help" data-shortcut="?"/>
-    </aside>
-</footer>
-</body>
-</html>
