--- ac1/htmlcov/status.json
@@ -1,1 +1,0 @@
-{"format":2,"version":"7.3.1","globals":"54f227e516abe37c03f4bb6295bcf432","files":{"d_5f5a17c013354698___init___py":{"hash":"3c77fc9ef7f887ac2508d4109cf92472","index":{"nums":[0,1,0,0,0,0,0,0],"html_filename":"d_5f5a17c013354698___init___py.html","relative_filename":"app\\__init__.py"}},"d_5f5a17c013354698_game_plat_py":{"hash":"55adf1614c4ac6fa3414f1fed9d2f9d2","index":{"nums":[0,1,56,0,4,0,0,0],"html_filename":"d_5f5a17c013354698_game_plat_py.html","relative_filename":"app\\game_plat.py"}},"d_a44f0ac069e85531___init___py":{"hash":"3c77fc9ef7f887ac2508d4109cf92472","index":{"nums":[0,1,0,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531___init___py.html","relative_filename":"tests\\__init__.py"}},"d_a44f0ac069e85531_test_usr_1_fail_py":{"hash":"4bce0d389ed96f238e62fdc806398339","index":{"nums":[0,1,2,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_1_fail_py.html","relative_filename":"tests\\test_usr_1_fail.py"}},"d_a44f0ac069e85531_test_usr_1_score_to_cripto_py":{"hash":"7b3b9ee4d0b526a1c7d72c7449589d98","index":{"nums":[0,1,4,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_1_score_to_cripto_py.html","relative_filename":"tests\\test_usr_1_score_to_cripto.py"}},"d_a44f0ac069e85531_test_usr_2_fail_py":{"hash":"d5015f198afd188ab2e1408df86fdcfd","index":{"nums":[0,1,2,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_2_fail_py.html","relative_filename":"tests\\test_usr_2_fail.py"}},"d_a44f0ac069e85531_test_usr_2_premium_users_py":{"hash":"4c3b5ce4a797e813b1d92ac150df8be3","index":{"nums":[0,1,5,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_2_premium_users_py.html","relative_filename":"tests\\test_usr_2_premium_users.py"}},"d_a44f0ac069e85531_test_usr_3_course_score_py":{"hash":"597fba09a8ab1b4ff93df395aa1be5e5","index":{"nums":[0,1,5,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_3_course_score_py.html","relative_filename":"tests\\test_usr_3_course_score.py"}},"d_a44f0ac069e85531_test_usr_3_fail_py":{"hash":"80d350144ab0e6feae9d443791ebb959","index":{"nums":[0,1,2,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_3_fail_py.html","relative_filename":"tests\\test_usr_3_fail.py"}},"d_a44f0ac069e85531_test_usr_4_course_bonus_py":{"hash":"64d0947403e258990b30301b6481a722","index":{"nums":[0,1,9,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_4_course_bonus_py.html","relative_filename":"tests\\test_usr_4_course_bonus.py"}},"d_a44f0ac069e85531_test_usr_4_fail_py":{"hash":"6ff87dc65c7eee1031a48ea1f32ae3bf","index":{"nums":[0,1,2,0,0,0,0,0],"html_filename":"d_a44f0ac069e85531_test_usr_4_fail_py.html","relative_filename":"tests\\test_usr_4_fail.py"}}}}
