--- ac1/tests/test_usr_1_score_to_cripto.py
@@ -1,7 +1,0 @@
-from app.game_plat import CoursePlatform
-
-
-def testPassStudentScoreToCripto():
-    platform = CoursePlatform()
-    assert platform.ScoreToCripto(
-        3, 5) == 2, "O número de moedas especificadas não passaram no teste!"
