--- ac1/tests/test_usr_2_fail.py
@@ -1,8 +1,0 @@
-from app.game_plat import CoursePlatform
-
-
-""" def testFailIdUsers():
-    platform = CoursePlatform()
-    platform.RegisterUser()
-    assert platform.FindPremiumUsers(), "Não existe usuários premiuns"
- """
