--- ac1/tests/test_usr_2_premium_users.py
@@ -1,7 +1,0 @@
-from app.game_plat import CoursePlatform
-
-
-def testPassIdUsers():
-    platform = CoursePlatform()
-    platform.RegisterUser(plan='Premium')
-    assert platform.FindPremiumUsers(), "Não existe usuários premiuns"
