--- ac1/tests/test_usr_3_fail.py
@@ -1,9 +1,0 @@
-from app.game_plat import CoursePlatform
-
-
-""" def testFailScoreToCourse():
-    platform = CoursePlatform()
-    platform.RegisterUser()
-    assert platform.ScoreToCourse(
-        ruler_number=-1), "O Usuário não possui curso ou a quantidade \
-            de cursos é superior/inferior que o ruler_number" """
