--- ac1/tests/test_usr_4_fail.py
@@ -1,6 +1,0 @@
-from app.game_plat import CoursePlatform
-
-
-""" def testFailScoreToCourse():
-    platform = CoursePlatform()
-    assert platform.Bonus(), 'Não há usuários ativos no forum' """
