+++ studies/tests/test_entre_16_e_18.py
@@ -1,0 +1,5 @@
+from app.main import Contratar
+
+def testMenor():
+    contratar = Contratar(idade = 17)
+    assert contratar.ContratarPessoa() == True, 'Idade não está entre 16 e 18 anos'
