+++ studies/tests/test_maior_55.py
@@ -1,0 +1,5 @@
+from app.main import Contratar
+
+def testMenor():
+    contratar = Contratar(idade = 56)
+    assert contratar.ContratarPessoa() == False, 'Idade maior que 56'
