+++ studies/tests/test_menor_16.py
@@ -1,0 +1,5 @@
+from app.main import Contratar
+
+def testMenor():
+    contratar = Contratar(idade = 15)
+    assert contratar.ContratarPessoa() == False, 'Idade inferior a 16 anos'
