+++ studies/tests/test_usr_1_fail.py
@@ -1,0 +1,7 @@
+from app.main import CoursePlatform
+
+""" def testFailStudentScoreToCripto():
+    platform = CoursePlatform()
+    assert platform.ScoreToCripto(
+        5, 3) == 2, "O número de moedas especificadas não passaram no teste!"
+ """
