+++ studies/tests/test_usr_2_fail.py
@@ -1,0 +1,8 @@
+from app.main import CoursePlatform
+
+
+""" def testFailIdUsers():
+    platform = CoursePlatform()
+    platform.RegisterUser()
+    assert platform.FindPremiumUsers(), "Não existe usuários premiuns"
+ """
