+++ studies/tests/test_usr_2_premium_users.py
@@ -1,0 +1,7 @@
+from app.main import CoursePlatform
+
+
+def testPassIdUsers():
+    platform = CoursePlatform()
+    platform.RegisterUser(plan='Premium')
+    assert platform.FindPremiumUsers(), "Não existe usuários premiuns"
