+++ studies/tests/test_usr_4_fail.py
@@ -1,0 +1,6 @@
+from app.main import CoursePlatform
+
+
+""" def testFailScoreToCourse():
+    platform = CoursePlatform()
+    assert platform.Bonus(), 'Não há usuários ativos no forum' """
